# docs/api/openapi.yaml
...
openapi: 3.1.0
info:
  title: go-proxy-manager admin API
  version: 0.1.0
  summary: 'REST API for the go-proxy-manager control plane: config CRUD, settings, backup/restore, DNS
    sync, Ingress discovery, and admin auth.'
  description: 'Covers the admin control-plane surface only (port 8081 by default), not the public data
    plane that serves proxied traffic. Every non-2xx response uses the uniform `{"error": "..."}` shape
    (see #/components/schemas/Error), except the plain-text/HTML auth endpoints noted individually below.


    Authentication: an admin session cookie (`sessionCookie`) or a scoped API bearer token (`bearerAuth`)
    - either satisfies every operation below unless noted. A session''s access is governed by its role
    (`user` for GET /api/me only; `admin` for everything under /api/); a token''s access is additionally
    governed by its granted scopes, shown per-operation as `x-scope` (`admin` satisfies every scope check;
    a concrete `<plural>:write` implies `<plural>:read`; `*:read`/`*:write` must be granted literally
    - a resource-scoped grant does not satisfy a wildcard requirement). Mutating session requests additionally
    require the `X-CSRF-Token` header (double-submit against the value GET /api/me returns); bearer requests
    never need it.


    HA: on an instance running GPM_HA_ROLE=follower, every write (any POST/PUT/DELETE under /api/) answers
    503 rather than being applied; reads are unaffected. See docs/ha.md.'
  license:
...
servers:
...
security:
- sessionCookie: []
- bearerAuth: []
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: gpm_session
      description: Admin browser session, issued by /auth/login (OIDC) or /auth/local (bcrypt). Cookie
        is HttpOnly, SameSite=Lax, and named __Host-gpm_session when GPM_COOKIE_SECURE=1 (default). A
        session principal has whatever access its role (admin/user) grants and is unaffected by API-token
        scopes. Mutating requests also need the X-CSRF-Token header.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: gpm_<random>
      description: A scoped API token (config kind api-tokens), created via PUT /api/api-tokens/{name}.
        Every token principal is admin-role by construction; real authorization is per-route scope checking
        (see each operation's x-scope). No CSRF header is used or required.
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message.
      required:
      - error
      description: Uniform error shape returned by every endpoint on a non-2xx response.
...
    Principal:
      type: object
      properties:
        Subject:
          type: string
          description: Stable IdP/local subject identifier. No json tag on this Go field, so the key is
            capitalized as shown.
        Email:
          type: string
        Name:
          type: string
        Role:
          type: string
          description: admin | user.
          enum:
          - admin
          - user
        IdP:
          type: string
          description: Name of the IdentityProvider that authenticated this session; empty for local login.
        csrfToken:
          type: string
          description: Required back as the X-CSRF-Token header on mutating session requests. Empty for
            a token principal.
        isToken:
          type: boolean
          description: True when authenticated by an API token rather than a session cookie.
        scopes:
          type: array
          items:
            type: string
          description: The token's granted scopes. Empty for a session principal (governed by role alone).
      description: The authenticated caller, as returned by GET /api/me.
...
  responses:
...
    Unauthorized:
      description: No valid session cookie or bearer token.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated, but the caller's role or API-token scope does not permit this action.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /api/me:
    get:
      summary: Current authenticated principal
      description: Proves the session/role gate and hands the SPA its CSRF token. Works for a session
        OR a bearer token.
      tags:
      - Auth
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Principal'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/openapi.yaml:
    get:
      summary: This OpenAPI 3.1 specification
      description: Same auth as GET /api/me (any authenticated session or token; no particular role or
        scope beyond authentication).
      tags:
      - Meta
      responses:
        '200':
          description: OK
          content:
            application/yaml:
              schema:
                type: string
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /auth/callback:
    get:
      summary: OIDC authorization-code callback
      description: Verifies state against the login cookie, exchanges the code, establishes the session,
        and redirects to the original return path.
      tags:
      - Auth
      parameters:
      - name: state
        in: query
        required: true
        schema:
          type: string
      - name: code
        in: query
        required: true
        schema:
          type: string
      - name: error
        in: query
        required: false
        schema:
          type: string
        description: Set by the IdP on a failed/denied authorization.
      responses:
        '302':
          description: Redirect to the original return path, with the session cookie set.
        '400':
          description: Missing state/code, or the state does not match the login cookie.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: The IdP reported an error, or the code exchange / identity checks failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
  /auth/local:
    post:
      summary: Local username/password login
      description: 'Same-origin only (CSRF belt: no session exists yet for a token-based defense). Rate-limited
        per client IP.'
      tags:
      - Auth
      responses:
        '302':
          description: Redirect to the return path, with the session cookie set.
        '403':
          description: Cross-origin request blocked (failed the same-origin guard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many recent failed attempts from this client IP.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Bad credentials, or local login is disabled/not configured.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                return:
                  type: string
                  description: Same-site path to redirect to on success.
  /auth/login:
    get:
      summary: Start (or choose) an admin login flow
      description: With ?idp=<name>, redirects straight to that IdentityProvider's OIDC flow. With none,
        renders an HTML chooser (SSO providers + local login form), unless exactly one SSO provider is
        configured and local login is off, in which case it redirects immediately (?select=1 forces the
        chooser, used by logout).
      tags:
      - Auth
      parameters:
      - name: idp
        in: query
        required: false
        schema:
          type: string
        description: IdentityProvider name to start an OIDC redirect immediately.
      - name: return
        in: query
        required: false
        schema:
          type: string
        description: Same-site path to return to after login. Anything else (open-redirect shapes) is
          discarded in favour of "/".
      - name: select
        in: query
        required: false
        schema:
          type: string
        description: Any non-empty value forces the chooser page even with a single SSO provider.
      responses:
        '200':
          description: HTML login chooser.
          content:
            text/html:
              schema:
                type: string
        '302':
          description: Redirect to the IdP's authorization endpoint.
        '502':
          description: The IdP could not be reached to start the flow.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security: []
  /auth/logout:
    post:
      summary: End the current admin session
      description: Same-origin only. Redirects to the login chooser (?select=1) rather than bouncing straight
        back into a still-live IdP session.
      tags:
      - Auth
      responses:
        '302':
          description: Redirect to /auth/login?select=1, with the session cookie cleared.
        '403':
          description: Cross-origin request blocked (failed the same-origin guard).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /debug/pprof/:
    get:
      summary: pprof index
      description: Only mounted when the instance is started with GPM_PPROF=1 (-pprof); a 404 here means
        profiling is off, not that this route doesn't exist. Gated by admin role, the same-origin guard,
        AND (unlike every other admin route) an explicit `admin` API-token scope check - a resource-scoped
        token gets 403, because a heap dump or /debug/pprof/cmdline carries resolved backend credentials
        in cleartext. `go tool pprof <url>` direct remote-attach does not work (no cookie, and its /debug/pprof/symbol
        POST fails CSRF); download the profile with curl -b and run pprof against the local file.
      tags:
      - Profiling (opt-in)
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      x-scope: admin
  /debug/pprof/cmdline:
    get:
      summary: Process command line
      description: Only mounted when the instance is started with GPM_PPROF=1 (-pprof); a 404 here means
        profiling is off, not that this route doesn't exist. Gated by admin role, the same-origin guard,
        AND (unlike every other admin route) an explicit `admin` API-token scope check - a resource-scoped
        token gets 403, because a heap dump or /debug/pprof/cmdline carries resolved backend credentials
        in cleartext. `go tool pprof <url>` direct remote-attach does not work (no cookie, and its /debug/pprof/symbol
        POST fails CSRF); download the profile with curl -b and run pprof against the local file.
      tags:
      - Profiling (opt-in)
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      x-scope: admin
  /debug/pprof/profile:
    get:
      summary: CPU profile (?seconds=N)
      description: Only mounted when the instance is started with GPM_PPROF=1 (-pprof); a 404 here means
        profiling is off, not that this route doesn't exist. Gated by admin role, the same-origin guard,
        AND (unlike every other admin route) an explicit `admin` API-token scope check - a resource-scoped
        token gets 403, because a heap dump or /debug/pprof/cmdline carries resolved backend credentials
        in cleartext. `go tool pprof <url>` direct remote-attach does not work (no cookie, and its /debug/pprof/symbol
        POST fails CSRF); download the profile with curl -b and run pprof against the local file.
      tags:
      - Profiling (opt-in)
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      x-scope: admin
  /debug/pprof/symbol:
    get:
      summary: Symbol lookup (used by go tool pprof)
      description: Only mounted when the instance is started with GPM_PPROF=1 (-pprof); a 404 here means
        profiling is off, not that this route doesn't exist. Gated by admin role, the same-origin guard,
        AND (unlike every other admin route) an explicit `admin` API-token scope check - a resource-scoped
        token gets 403, because a heap dump or /debug/pprof/cmdline carries resolved backend credentials
        in cleartext. `go tool pprof <url>` direct remote-attach does not work (no cookie, and its /debug/pprof/symbol
        POST fails CSRF); download the profile with curl -b and run pprof against the local file.
      tags:
      - Profiling (opt-in)
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      x-scope: admin
  /debug/pprof/trace:
    get:
      summary: Execution trace (?seconds=N)
      description: Only mounted when the instance is started with GPM_PPROF=1 (-pprof); a 404 here means
        profiling is off, not that this route doesn't exist. Gated by admin role, the same-origin guard,
        AND (unlike every other admin route) an explicit `admin` API-token scope check - a resource-scoped
        token gets 403, because a heap dump or /debug/pprof/cmdline carries resolved backend credentials
        in cleartext. `go tool pprof <url>` direct remote-attach does not work (no cookie, and its /debug/pprof/symbol
        POST fails CSRF); download the profile with curl -b and run pprof against the local file.
      tags:
      - Profiling (opt-in)
      responses:
        '200':
          description: OK
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
      x-scope: admin
...